requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
mysqlclient==2.1.0
protobuf==3.19.4
PyJWT==1.7.1
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
Werkzeug==2.0.3
